+++ analyzers/src/SonarAnalyzer.CFG/ShimLayer/NullableAnnotation.cs
@@ -1,0 +1,23 @@
+﻿// Copyright (c) Tunnel Vision Laboratories, LLC. All Rights Reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+namespace StyleCop.Analyzers.Lightup
+{
+    public enum NullableAnnotation
+    {
+        /// <summary>
+        /// The expression has not been analyzed, or the syntax is not an expression (such as a statement).
+        /// </summary>
+        None = 0,
+
+        /// <summary>
+        /// The expression is not annotated (does not have a ?).
+        /// </summary>
+        NotAnnotated = 1,
+
+        /// <summary>
+        /// The expression is annotated (does have a ?).
+        /// </summary>
+        Annotated = 2,
+    }
+}
