+++ analyzers/src/SonarAnalyzer.CFG/ShimLayer/TypeInfoExtensions.cs
@@ -42,63 +42,4 @@ namespace StyleCop.Analyzers.Lightup
             return expression.Compile();
         }
     }
-
-    public readonly record struct NullabilityInfo
-    {
-        public NullabilityInfo(NullableAnnotation annotation, NullableFlowState flowState)
-        {
-            Annotation = annotation;
-            FlowState = flowState;
-        }
-
-        /// <summary>
-        /// The nullable annotation of the expression represented by the syntax node. This represents
-        /// the nullability of expressions that can be assigned to this expression, if this expression
-        /// can be used as an lvalue.
-        /// </summary>
-        public NullableAnnotation Annotation { get; }
-
-        /// <summary>
-        /// The nullable flow state of the expression represented by the syntax node. This represents
-        /// the compiler's understanding of whether this expression can currently contain null, if
-        /// this expression can be used as an rvalue.
-        /// </summary>
-        public NullableFlowState FlowState { get; }
-    }
-
-    public enum NullableAnnotation
-    {
-        /// <summary>
-        /// The expression has not been analyzed, or the syntax is not an expression (such as a statement).
-        /// </summary>
-        None = 0,
-
-        /// <summary>
-        /// The expression is not annotated (does not have a ?).
-        /// </summary>
-        NotAnnotated = 1,
-
-        /// <summary>
-        /// The expression is annotated (does have a ?).
-        /// </summary>
-        Annotated = 2,
-    }
-
-    public enum NullableFlowState
-    {
-        /// <summary>
-        /// Syntax is not an expression, or was not analyzed.
-        /// </summary>
-        None = 0,
-
-        /// <summary>
-        /// Expression is not null.
-        /// </summary>
-        NotNull = 1,
-
-        /// <summary>
-        /// Expression may be null.
-        /// </summary>
-        MaybeNull = 2,
-    }
 }
